# Add SnoopPrecompile