--- a/Project.toml
+++ b/Project.toml
@@ -7,6 +7,7 @@ version = "1.9.5"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Mmap = "a63ad114-7e13-5084-954f-fe012c677804"
 Parsers = "69de0a69-1ddd-5017-9359-2bf0b02dc9f0"
+SnoopPrecompile = "66db9d55-30c0-4569-8b51-7e840670fc0c"
 StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
@@ -14,6 +15,7 @@ UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 Parsers = "0.3, 1, 2"
 StructTypes = "1.5"
 julia = "1"
+SnoopPrecompile = "1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"