The small dependency SnoopPrecompile helps precompilation, while itself having
little impact on latency.
With this change, JSON3 precompiles significantly better.

This change increases time spent on `using`, but decreases inference time for any of the workloads in the `workload.jl` file.

Here are timings for a script which loads JSON3 and executs the same workload in `workload.jl` file on master vs this PR:

```
# Master
Benchmark 1: julia --project=. --startup=no /tmp/foo.jl
  Time (mean ± σ):      3.531 s ±  0.065 s    [User: 3.537 s, System: 0.580 s]
  Range (min … max):    3.453 s …  3.697 s    10 runs

# PR
Benchmark 1: julia --project=. --startup=no /tmp/foo.jl
  Time (mean ± σ):      2.149 s ±  0.031 s    [User: 2.195 s, System: 0.554 s]
  Range (min … max):    2.099 s …  2.206 s    10 runs
```

@quinnj I'm not sure the workload covers all the common use-cases of JSON3, i.e. whether the workload ends up calling all the methods a user would typically call.