@jakobnissen 2022-08-30
The workload from #210 is cut from 1.94s to 1.20s after this change.