@jakobnissen 2022-10-17
@quinnj is it possible to merge this? The precise SnoopPrecompile statements can easily be updated later.